Refresh happy-dom option selection cache
