+++ packages/ariakit-test/src/click.jsdom.test.ts
@@ -1,0 +1,58 @@
+// @vitest-environment jsdom
+
+import { afterEach, expect, test } from "vitest";
+import { click, q } from "./index.ts";
+
+afterEach(() => {
+  document.body.innerHTML = "";
+});
+
+function setupMultiSelect() {
+  document.body.innerHTML = `
+    <label for="fruits">Fruits</label>
+    <select id="fruits" multiple size="3">
+      <option>Apple</option>
+      <option selected>Banana</option>
+      <option>Cherry</option>
+    </select>
+  `;
+}
+
+function getSelect() {
+  return q.listbox.ensure("Fruits") as HTMLSelectElement;
+}
+
+function getSelectedValues() {
+  return Array.from(getSelect().options)
+    .filter((option) => option.selected)
+    .map((option) => option.value);
+}
+
+function getDefaultSelectedValues() {
+  return Array.from(getSelect().options)
+    .filter((option) => option.defaultSelected)
+    .map((option) => option.value);
+}
+
+test("clicking options does not change their default selectedness", async () => {
+  setupMultiSelect();
+
+  await click(q.option.ensure("Apple"));
+  await click(q.option.ensure("Cherry"), { ctrlKey: true });
+  await click(q.option.ensure("Apple"), { ctrlKey: true });
+
+  expect(getSelectedValues()).toEqual(["Cherry"]);
+  expect(getDefaultSelectedValues()).toEqual(["Banana"]);
+  expect(q.option.ensure("Apple")).not.toHaveAttribute("selected");
+  expect(q.option.ensure("Banana")).toHaveAttribute("selected");
+  expect(q.option.ensure("Cherry")).not.toHaveAttribute("selected");
+});
+
+test("anchorless shift-click anchors at the first selected option", async () => {
+  setupMultiSelect();
+
+  await click(q.option.ensure("Cherry"), { shiftKey: true });
+
+  expect(getSelectedValues()).toEqual(["Banana", "Cherry"]);
+  expect(getDefaultSelectedValues()).toEqual(["Banana"]);
+});
